requirements.txt
pandas
numpy
python-dotenv
requests
xmltodict
opencv-python
pillow